CORE-8536 Fix :each_hour_form method for Utils::Time
It appeared billing statistics is not being calculated for not trivial
zones (like IST). This pull-request fixes it with adding .utc to each
iterator step. For zones like ISZ time differs from UTS in 5:30 (not
numerical amount). That is why
>> Time.new(2014, 2, 17,16,25).utc.beginning_of_hour
=> 2014-02-17 10:00:00 UTC
but
>> Time.new(2014, 2, 17,16,35).utc.beginning_of_hour
=> 2014-02-17 11:00:00 UTC

So +10 minutes gives us +1 hour and broken statistics. Because
user_hourly_stats are created and their :created_at = 10:30, but
stat_time = 11:00.

It was broken in 5.2 in CORE-8033 and 68ad6a2 commit
(https://bitbucket.org/onappcore/onapp-utils/commits/68ad6a2490fa84099d68c5c77cdeadb1ffc54546#Llib/utils/time.rbT38)

This fix is going to be cherry-picked into 5.3.* versions of onapp-utils